remove max=0 meaning no limit

+++ streams.h
@@ -40,28 +40,23 @@ extern int stream_length(Stream *);
 
 extern void enable_stream_exceptions();
 extern void disable_stream_exceptions();
-extern int stream_alloc_maximum;
+extern size_t stream_alloc_maximum;
 extern Exception stream_too_big;
 /*
  * Calls to enable_stream_exceptions() and disable_stream_exceptions()
  * must be paired and nest properly.
  *
- * If (1) enable_stream_exceptions() is in effect, and
- * (2) stream_alloc_maximum is set to a positive value,
- * then, upon any attempt to grow a stream beyond
- * stream_alloc_maximum bytes, a stream_too_big exception
- * will be raised.
+ * If enable_stream_exceptions() is in effect, then, upon any
+ * attempt to grow a stream beyond stream_alloc_maximum bytes,
+ * a stream_too_big exception will be raised.
  */
 
 #endif
 
 /* 
  * $Log: streams.h,v $
- * Revision 1.6  2010/04/22 21:23:24  wrog
- * whitespace fixes
- *
- * Revision 1.5  2010/03/30 22:13:22  wrog
- * Added stream exception API to catch mymalloc failures
+ * Revision 1.7  2010/04/23 05:10:55  wrog
+ * remove max=0 meaning no limit
  *
  * Revision 1.4  2006/12/06 23:57:51  wrog
  * New INPUT_APPLY_BACKSPACE option to process backspace/delete characters on nonbinary connections (patch 1571939)
